% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\docType{data}
\name{API_KEY_GOOGLE_Gemini}
\alias{API_KEY_GOOGLE_Gemini}
\title{API KEY GOOGLE Gemini}
\format{
An object of class \code{character} of length 1.
}
\usage{
API_KEY_GOOGLE_Gemini
}
\description{
API for GOOGLE Gemini
}
\keyword{datasets}
