% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_4_Population_Structure.R
\name{Page_4_Population_Structure_UI}
\alias{Page_4_Population_Structure_UI}
\title{Page_4_Population_Structure_UI}
\usage{
Page_4_Population_Structure_UI()
}
\description{
Page_4_Population_Structure_UI
}
