% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page_7_Core_Collection.R
\name{Page_7_Core_Collection_Server}
\alias{Page_7_Core_Collection_Server}
\title{Page_7_Core_Collection_Server}
\usage{
Page_7_Core_Collection_Server(input, output, session)
}
\description{
Page_7_Core_Collection_Server
}
